+++ magda-gateway/src/buildApp.ts
@@ -43,6 +43,9 @@ type Config = {
     proxyRoutesJson: {
         [localRoute: string]: ProxyTarget;
     };
+    webProxyRoutesJson: {
+        [localRoute: string]: string;
+    };
     helmetJson: string;
     cspJson: string;
     corsJson: string;
@@ -197,6 +200,13 @@ export default function buildApp(config: Config) {
             tenantMode
         })
     );
+
+    if (config.webProxyRoutesJson) {
+        _.forEach(config.webProxyRoutesJson, (value: string, key: string) => {
+            app.use("/" + key, createGenericProxy(value, tenantMode));
+        });
+    }
+
     app.use("/preview-map", createGenericProxy(config.previewMap, tenantMode));
 
     if (config.enableCkanRedirection) {